requirements.txt
aniso8601==8.0.0
click==7.1.2
flake8==3.8.3
Flask==1.1.2
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==7.2.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
six==1.15.0
Werkzeug==1.0.1
